Update attribute evaluation and generation process
Previously if the attributes were not evaluated and the user tried to set an attribute by code or via blueprints the attribute evaluation was triggered and then the attribute was set. This could lead to strange side effects if for example immediately after setting an attribute a generate call was triggered (which would internally cancel the attribute evaluation and therefore the attribute would never be set).

We simplify the whole process now by allowing the user to set attributes before they were evaluated once and just treating them as user set attributes for the evaluation.